requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
idna==3.6
imageio==2.36.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==2.2.1
openapi-codec==1.3.2
pillow==10.4.0
proglog==0.1.10
psycopg==3.2.9
psycopg2==2.9.10
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.25.1
setuptools==75.6.0
simplejson==3.19.2
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
